requirements.txt
google-genai>=0.2.0
flask>=2.3.0
flask-cors>=4.0.0
